Update feature complete
Update feature only updates task if it is actually different from the original task. Bug: If a second task is added after the first task, and the first task is modified, when diplaying the tasks, the task that was modified is listed after the second task although it was created first. Fix: Order tasks by id instead of date.